requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
email-validator>=2.1.1
aiofiles>=23.2.0
httpx>=0.27.0
groq>=0.4.2
python-multipart>=0.0.9
pdfplumber>=0.11.0
pillow>=10.0.0
